backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
pymongo==4.6.0
redis==5.0.1
kafka-python==2.0.2
scikit-learn==1.3.2
prophet==1.1.5
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
boto3==1.29.7
azure-storage-blob==12.19.0
google-cloud-storage==2.10.0
cryptography==41.0.7
pyjwt==2.8.0
passlib==1.7.4
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.9.1
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3
prometheus-client==0.19.0
python-socketio==5.10.0
websockets==12.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
scipy==1.11.4
